+++ lib/active_interaction.rb
@@ -42,5 +42,5 @@ I18n.backend.load_translations(
 #
 # @since 1.0.0
 #
-# @version 1.0.0
+# @version 1.0.1
 module ActiveInteraction end