+++ CHANGELOG.md
@@ -6,6 +6,13 @@
   was required.
 - Speed up many filters by caching class constants.
 
+# [1.0.1][] (2014-02-04)
+
+- Short circuit `valid?` after successfully running an interaction.
+- Fix a bug that prevented merging interpolated symbolic errors.
+- Use `:invalid_type` instead of `:invalid` as I18n key for type errors.
+- Fix a bug that skipped setting up accessors for imported filters.
+
 # [1.0.0][] (2014-01-21)
 
 - **Replace `Filters` with a hash.** To iterate over `Filter` objects, use
@@ -124,7 +131,8 @@
 
 - Initial release.
 
-  [master]: https://github.com/orgsync/active_interaction/compare/v1.0.0...master
+  [master]: https://github.com/orgsync/active_interaction/compare/v1.0.1...master
+  [1.0.0]: https://github.com/orgsync/active_interaction/compare/v1.0.0...v1.0.1
   [1.0.0]: https://github.com/orgsync/active_interaction/compare/v0.10.2...v1.0.0
   [0.10.2]: https://github.com/orgsync/active_interaction/compare/v0.10.1...v0.10.2
   [0.10.1]: https://github.com/orgsync/active_interaction/compare/v0.10.0...v0.10.1